PB-3694 - Encrypted Description feedback
PB-3695 - [SHOULD] When editing the Description, it should include an icon (with tooltip on hover) indicating whether it's encrypted or not.
PB-3696 - [MUST] When editing an encrypted description and when there is a decryption error, we display a message in italic, "decryption failed, click here to retry".
PB-3697 - [SHOULD] When editing, the space between Description edit and buttons needs to be the same as comment section